+++ src/netevt_select.c
@@ -139,7 +139,7 @@ static int netevt_select_wait(struct netevt_impl* impl, netevt obj, int timeout,
 
         if(nb > nb_events)
         {
-          return nb;
+          return (int)nb;
         }
 
         if(i == 0)
